Tidy up a bit
